refactor(ale_claw): group provider/inference modules into harness/inference/ subpackage
Move model_config, thinking, cache_policy, unified_loop, helper_runtime,
_message_shapes into an inference/ subpackage (the 'talking to the model'
layer, mirroring pi's separate ai package). Behavior-preserving: git mv +
import-path updates incl. the unified_loop side-effect registration import.
928 passed, 3 skipped.
