+++ ale_run/agents/ale_claw/harness/agent_loop_helpers.py
@@ -12,7 +12,7 @@ import base64 as _base64
 import re
 from typing import TYPE_CHECKING, Any, Dict, List
 
-from ._message_shapes import _image_url_block
+from .inference._message_shapes import _image_url_block
 from .image_sanitization import (
     DEFAULT_LIMITS as _IMAGE_DEFAULT_LIMITS,
     sanitize_raw_image_bytes as _sanitize_raw_image_bytes,
