+++ ale_run/agents/ale_claw/harness/inference/thinking.py
