+++ ale_run/agents/ale_claw/harness/subagent/subagent_session.py
@@ -34,7 +34,7 @@ import os
 from pathlib import Path
 from typing import Any
 
-from ..model_config import ResolvedModel, resolve_model
+from ..inference.model_config import ResolvedModel, resolve_model
 from agent.tools.base import BaseTool
 
 from ..context.context import (
